+++ test/aggregation-examples/trend-analysis/largest-graph-network.test.ts
@@ -1,3 +1,5 @@
+import { aggregate } from "../../support";
+
 /**
  * Your organisation wants to know the best targets for a new marketing campaign based on a social network database similar to Twitter.
  * You want to search the collection of social network users, each holding a user's name and the names of other people who follow them.
@@ -23,7 +25,7 @@ describe("Largest Graph Network", () => {
     // For each social network user, graph traverse their 'followed_by' list of people
     {
       $graphLookup: {
-        from: "users",
+        from: users,
         startWith: "$followed_by",
         connectFromField: "followed_by",
         connectToField: "name",
@@ -59,77 +61,81 @@ describe("Largest Graph Network", () => {
       $sort: {
         network_reach: -1
       }
+    },
+    // ensure stable array fields
+    {
+      $addFields: {
+        extended_connections: {
+          $sortArray: { input: "$extended_connections", sortBy: { _id: 1 } }
+        }
+      }
     }
   ];
 
-  it("passes", () => {
-    expect(true).toBe(true);
+  it("return ten documents should be returned, corresponding to the original ten source social network users, with each one including a count of the user's network reach, and the names of their extended connections, sorted by the user with the most extensive network reach first", () => {
+    expect(aggregate(users, pipeline)).toEqual([
+      {
+        name: "Carol",
+        network_reach: 8,
+        extended_connections: [
+          "Carl",
+          "David",
+          "Fiona",
+          "Helen",
+          "Janet",
+          "Paul",
+          "Sarah",
+          "Toni"
+        ]
+      },
+      {
+        name: "Sarah",
+        network_reach: 6,
+        extended_connections: [
+          "Carl",
+          "David",
+          "Fiona",
+          "Janet",
+          "Paul",
+          "Toni"
+        ]
+      },
+      {
+        name: "Carl",
+        network_reach: 5,
+        extended_connections: ["David", "Fiona", "Janet", "Paul", "Toni"]
+      },
+      {
+        name: "Fiona",
+        network_reach: 4,
+        extended_connections: ["David", "Janet", "Paul", "Toni"]
+      },
+      {
+        name: "David",
+        network_reach: 3,
+        extended_connections: ["Janet", "Paul", "Toni"]
+      },
+      {
+        name: "Bob",
+        network_reach: 3,
+        extended_connections: ["Janet", "Paul", "Toni"]
+      },
+      {
+        name: "Janet",
+        network_reach: 2,
+        extended_connections: ["Paul", "Toni"]
+      },
+      {
+        name: "Toni",
+        network_reach: 1,
+        extended_connections: ["Paul"]
+      },
+      {
+        name: "Helen",
+        network_reach: 1,
+        extended_connections: ["Paul"]
+      },
+      { name: "Paul", network_reach: 0, extended_connections: [] }
+    ]);
   });
-
-  // it("return ten documents should be returned, corresponding to the original ten source social network users, with each one including a count of the user's network reach, and the names of their extended connections, sorted by the user with the most extensive network reach first", () => {
-  //   expect(aggregate(users, pipeline, DEFAULT_OPTS)).not.toEqual([
-  //     {
-  //       name: "Carol",
-  //       network_reach: 8,
-  //       extended_connections: [
-  //         "David",
-  //         "Toni",
-  //         "Fiona",
-  //         "Sarah",
-  //         "Helen",
-  //         "Carl",
-  //         "Paul",
-  //         "Janet"
-  //       ]
-  //     },
-  //     {
-  //       name: "Sarah",
-  //       network_reach: 6,
-  //       extended_connections: [
-  //         "David",
-  //         "Toni",
-  //         "Fiona",
-  //         "Carl",
-  //         "Paul",
-  //         "Janet"
-  //       ]
-  //     },
-  //     {
-  //       name: "Carl",
-  //       network_reach: 5,
-  //       extended_connections: ["David", "Toni", "Fiona", "Paul", "Janet"]
-  //     },
-  //     {
-  //       name: "Fiona",
-  //       network_reach: 4,
-  //       extended_connections: ["David", "Toni", "Paul", "Janet"]
-  //     },
-  //     {
-  //       name: "David",
-  //       network_reach: 3,
-  //       extended_connections: ["Toni", "Paul", "Janet"]
-  //     },
-  //     {
-  //       name: "Bob",
-  //       network_reach: 3,
-  //       extended_connections: ["Toni", "Paul", "Janet"]
-  //     },
-  //     {
-  //       name: "Janet",
-  //       network_reach: 2,
-  //       extended_connections: ["Toni", "Paul"]
-  //     },
-  //     {
-  //       name: "Toni",
-  //       network_reach: 1,
-  //       extended_connections: ["Paul"]
-  //     },
-  //     {
-  //       name: "Helen",
-  //       network_reach: 1,
-  //       extended_connections: ["Paul"]
-  //     },
-  //     { name: "Paul", network_reach: 0, extended_connections: [] }
-  //   ]);
-  // });
 });